move admin to its own directory
